Final Submission

+++ HCTree.cpp
@@ -22,16 +22,24 @@ using namespace std;
  * nodes and the leaves vector.
  */
 HCTree::~HCTree(){
-  deleteAll(root);
+  
+  // if root exists (if there is any content in the file)
+  // then delete everything
+  if (root) {
+    deleteAll(root);
+  }
 }
 
 /**
  * This method builds the HUffman tree used to encode.
  */
 void HCTree::build(const vector<int>& freqs){
+
+  // make a priority queue to build the tree using Huffman algorithm
   std::priority_queue<HCNode*,std::vector<HCNode*>,HCNodePtrComp> pq;
+
+  // put symbols that appear in the file into the pq
   for (unsigned int i = 0; i < freqs.size(); i++) {
-    //HCNode * c0, c1, p;
     if(freqs.at(i) > 0){
       HCNode* p = new HCNode(freqs[i], i, 0, 0, 0);
       pq.push(p);
@@ -53,23 +61,31 @@ void HCTree::build(const vector<int>& freqs){
 
     unsigned int summedCount = bot1->count + bot2->count;
 
-    //byte emptySym = 0;
 
+    // create a parent node from the smallest nodes
     HCNode* sumNode = new HCNode(summedCount, bot2->symbol, bot1, bot2, 0);
     bot1->p = sumNode;
     bot2->p = sumNode;
     pq.push(sumNode);
   }
-  root = pq.top();
-  pq.pop();
+
+  // set root if file is not empty
+  if (!pq.empty()) {
+    root = pq.top();
+    pq.pop();
+  }
 }
+
+/**
+ * This method encodes individual bits and writes them to the output file
+ */
 
 void HCTree::encode(byte symbol, BitOutputStream& out) const{
   HCNode * currentNode; // node to traverse through tree
   HCNode * parentNode; // parent of currentNode
   std::stack<int> bits; // temp stack to hold encoded bits
   int bit; // bit in encoded sequence for symbol
-  unsigned int bits_size = 0;
+  unsigned int bits_size = 0; // number of bits in sequences for each symb
 
   // look in tree for symbol's node
   for (int i = 0; i < leaves.size(); i++) {
@@ -81,22 +97,23 @@ void HCTree::encode(byte symbol, BitOutputStream& out) const{
   // loop through from symbol node to root and find bit reverse bit
   // sequence
   if(currentNode->p){
-  while (currentNode->p) {
-    parentNode = currentNode->p;
-    // if right child, bit is 0
-    if (currentNode == parentNode->c0) {
-      bits.push(0);
-      bits_size++;
+    while (currentNode->p) {
+      parentNode = currentNode->p;
+      // if right child, bit is 0
+      if (currentNode == parentNode->c0) {
+        bits.push(0);
+        bits_size++;
+      }
+      // if left child, bit is 1
+      else {
+        bits.push(1);
+        bits_size++;
+      }
+      currentNode = parentNode;
     }
-    // if left child, bit is 1
-    else {
-      bits.push(1);
-      bits_size++;
-    }
-    currentNode = parentNode;
-  }
   }
 
+  // if only one node in tree, bit sequence is 1
   else if(root == currentNode){
     bits.push(1);
     bits_size++;
@@ -105,7 +122,7 @@ void HCTree::encode(byte symbol, BitOutputStream& out) const{
   for(int i=0; i<bits_size; i++){
     //cout << bits_size << endl;
     bit = bits.top();
-    cout << "Bits Printed: "<< bit << endl;
+    //cout << "Bits Printed: "<< bit << endl;
     out.writeBit(bit); // write to file's buffer as bit
     bits.pop();
   }
@@ -161,21 +178,33 @@ void HCTree::encode(byte symbol, ofstream& out) const{
 
 }
 
+/**
+ * This method decodes bits of data from the outfile of encode
+ */
+
 int HCTree::decode(BitInputStream& in) const{
   unsigned char readBit;
-  int readNum;
-  HCNode * currentNode = root;
+  int readNum; // number being read in from the input file
+  HCNode * currentNode = root; // set root as working node
+
+  // keep reading the bits and following tree 
+  // while currentNode has children
   while(currentNode->c0 && currentNode->c1){
-    readNum = in.readBit();
+    readNum = in.readBit(); // read in a byte from file
+
+    // break loop if end of file reached
     if(readNum == EOF){
       break;
     }
     //readBit = (unsigned char)readNum;
     //cout << "readBit: " << readNum << endl;
+
+    // if 0 read in, travel down right child
     if(readNum == 0){
       currentNode = currentNode->c0;
     }
 
+    // if 1 read in, travel down left child
     else{
       currentNode = currentNode->c1;
     }
@@ -184,10 +213,12 @@ int HCTree::decode(BitInputStream& in) const{
 //      break;
 //    }
 
+    // break loop if end of file is reached
     if(in.eof()){
       break;
     }
   }
+
   return currentNode->symbol;
   
 }
